requirements.txt
torch
torchvision
ultralytics>=8.0.0
chromadb
psycopg2-binary
pandas
matplotlib
seaborn
Pillow
moviepy
coremltools
transformers
numpy
python-dotenv
sentence-transformers